moving healthbar to components.ui package
